Rename file to match existing pattern
